+++ cricbuzz-microservices/MatchesDb/src/main/java/com/cricbuzz/MatchesDbApplication.java
@@ -1,0 +1,17 @@
+package com.cricbuzz;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+
+@SpringBootApplication
+@EnableEurekaClient
+public class MatchesDbApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MatchesDbApplication.class, args);
+	}
+
+}
+
