+++ cricbuzz-microservices/MatchesDb/src/main/java/com/cricbuzz/repository/MatchesRepository.java
@@ -1,0 +1,9 @@
+package com.cricbuzz.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.cricbuzz.model.Matches;
+
+public interface MatchesRepository extends MongoRepository<Matches, String> {
+
+}
