--- cricbuzz-microservices/MatchesDb/src/main/java/com/example/demo/MatchesDbApplication.java
@@ -1,17 +1,0 @@
-package com.example.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-
-@SpringBootApplication
-@EnableEurekaClient
-public class MatchesDbApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MatchesDbApplication.class, args);
-	}
-
-}
-
