--- cricbuzz-microservices/MatchesDb/src/main/java/com/example/demo/repository/MatchesRepository.java
@@ -1,9 +1,0 @@
-package com.example.demo.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.example.demo.model.Matches;
-
-public interface MatchesRepository extends MongoRepository<Matches, String> {
-
-}
