--- cricbuzz-microservices/MatchesDb/src/main/java/com/example/demo/swagger/SwaggerConfig.java
@@ -1,23 +1,0 @@
-package com.example.demo.swagger;
-import static springfox.documentation.builders.PathSelectors.regex;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@Configuration
-public class SwaggerConfig {
-
-	@Bean
-	public Docket productApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.example.demo.dao"))
-                .paths(regex("/match.*"))
-                .build();
-    }
-}
