+++ cricbuzz-microservices/PlayersDb/src/main/java/com/cricbuzz/PlayersDbApplication.java
@@ -1,0 +1,16 @@
+package com.cricbuzz;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class PlayersDbApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PlayersDbApplication.class, args);
+	}
+
+}
+
