+++ cricbuzz-microservices/PlayersDb/src/main/java/com/cricbuzz/repository/PlayersRepository.java
@@ -1,0 +1,8 @@
+package com.cricbuzz.repository;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.cricbuzz.model.Players;
+
+public interface PlayersRepository extends MongoRepository<Players, String> {
+}
