--- cricbuzz-microservices/PlayersDb/src/main/java/com/demo/PlayersDbApplication.java
@@ -1,16 +1,0 @@
-package com.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class PlayersDbApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PlayersDbApplication.class, args);
-	}
-
-}
-
