--- cricbuzz-microservices/PlayersDb/src/main/java/com/demo/repository/PlayersRepository.java
@@ -1,8 +1,0 @@
-package com.demo.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.demo.model.Players;
-
-public interface PlayersRepository extends MongoRepository<Players, String> {
-}
