--- cricbuzz-microservices/RankingService/src/main/java/com/RankingService/RankingServiceApplication.java
@@ -1,16 +1,0 @@
-package com.RankingService;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class RankingServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RankingServiceApplication.class, args);
-	}
-
-}
-
