--- cricbuzz-microservices/RankingService/src/main/java/com/RankingService/repository/RankingRepository.java
@@ -1,13 +1,0 @@
-package com.RankingService.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.RankingService.pojo.Rankings;
-
-public interface RankingRepository extends MongoRepository<Rankings, String>{
-
-	
-
-}
