+++ cricbuzz-microservices/RankingService/src/main/java/com/cricbuzz/RankingServiceApplication.java
@@ -1,0 +1,16 @@
+package com.cricbuzz;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+
+@SpringBootApplication
+@EnableEurekaClient
+public class RankingServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RankingServiceApplication.class, args);
+	}
+
+}
+
