+++ cricbuzz-microservices/RankingService/src/main/java/com/cricbuzz/repository/RankingRepository.java
@@ -1,0 +1,13 @@
+package com.cricbuzz.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.cricbuzz.model.Rankings;
+
+public interface RankingRepository extends MongoRepository<Rankings, String>{
+
+	
+
+}
