--- cricbuzz-microservices/UpComingMatches/src/main/java/com/upcomingmatches/UpcomingMatchesApplication.java
@@ -1,14 +1,0 @@
-package com.upcomingmatches;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UpcomingMatchesApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UpcomingMatchesApplication.class, args);
-	}
-
-}
-
