--- cricbuzz-microservices/UpComingMatches/src/main/java/com/upcomingmatches/repo/UpcomingMatchesRepo.java
@@ -1,11 +1,0 @@
-package com.upcomingmatches.repo;
-
-import java.util.List;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.upcomingmatches.model.UpcomingMatches;
-
-public interface UpcomingMatchesRepo extends MongoRepository<UpcomingMatches, String> {
-
-}
