--- cricbuzz-microservices/UpComingMatches/src/main/java/com/upcomingmatches/swagger/SwaggerConfig.java
@@ -1,47 +1,0 @@
-package com.upcomingmatches.swagger;
-
-import static springfox.documentation.builders.PathSelectors.regex;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@Configuration
-public class SwaggerConfig {
-	
-	@Bean
-    public Docket productApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.upcomingmatches"))
-                .paths(regex("/.*"))
-               
-                .build()
-                .apiInfo(metaInfo());
-    }
-
-    private ApiInfo metaInfo() {
-
-        ApiInfo apiInfo = new ApiInfo(
-                "Spring Boot Swagger Example API",
-                "Spring Boot Swagger Example API for Youtube",
-                "1.0",
-                "Terms of Service",
-                new Contact("Vaibhavi", "https://www.youtube.com",
-                        "xyz@gmail.com"),
-                "Apache License Version 0.0",
-                "https://www.apache.org/licesen.html"
-        );
-
-        return apiInfo;
-    }
-
-}
-
