--- cricbuzz-microservices/UpComingMatches/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-server.port=6000
-spring.data.mongodb.database=UpcomingMatches
-spring.data.mongodb.port=27017
-spring.security.user.name=cricbuzz
-spring.security.user.password=123
-eureka.client.healthcheck.enabled= true
-spring.application.name=upcomingMatches
-eureka.client.eureka-server-port=8761
