+++ cricbuzz-microservices/live-scores/src/main/java/com/cricbuzz/config/SwaggerConfig.java
@@ -14,12 +14,12 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 @Configuration
 public class SwaggerConfig {
 
-	@Bean
-	public Docket productApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com"))
-                .paths(regex("*"))
-                .build();
-    }
+	  @Bean
+	    public Docket productApi() {
+	        return new Docket(DocumentationType.SWAGGER_2)
+	                .select()
+	                .apis(RequestHandlerSelectors.basePackage("com.simer.livescores"))
+	                .paths(regex("/getscore.*"))
+	                .build();
+	    }
 }
