+++ cricbuzz-microservices/live-scores/src/main/java/com/cricbuzz/repository/ScoreRepository.java
@@ -1,0 +1,8 @@
+package com.cricbuzz.repository;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+import com.cricbuzz.model.LiveScores;
+
+public interface ScoreRepository extends MongoRepository<LiveScores, String>{
+
+}
