--- cricbuzz-microservices/live-scores/src/main/java/com/simer/livescores/config/SwaggerConfig.java
@@ -1,25 +1,0 @@
-package com.simer.livescores.config;
-
-import static springfox.documentation.builders.PathSelectors.regex;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@EnableSwagger2
-@Configuration
-public class SwaggerConfig {
-
-	  @Bean
-	    public Docket productApi() {
-	        return new Docket(DocumentationType.SWAGGER_2)
-	                .select()
-	                .apis(RequestHandlerSelectors.basePackage("com.simer.livescores"))
-	                .paths(regex("/getscore.*"))
-	                .build();
-	    }
-}
