--- cricbuzz-microservices/live-scores/src/main/java/com/simer/livescores/repository/ScoreRepository.java
@@ -1,8 +1,0 @@
-package com.simer.livescores.repository;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.simer.livescores.model.LiveScores;
-
-public interface ScoreRepository extends MongoRepository<LiveScores, String>{
-
-}
